+++ src/main/java/awtoz/awtoz/hearttransaction/exception/InvalidHeartAmountException.java
@@ -1,4 +1,4 @@
-package awtoz.awtoz.heart.exception;
+package awtoz.awtoz.hearttransaction.exception;
 
 public class InvalidHeartAmountException extends RuntimeException {
     public InvalidHeartAmountException(String message) {
