+++ src/main/java/awtoz/awtoz/hearttransaction/exception/InvalidHeartTransactionTypeException.java
@@ -1,4 +1,4 @@
-package awtoz.awtoz.heart.exception;
+package awtoz.awtoz.hearttransaction.exception;
 
 public class InvalidHeartTransactionTypeException extends RuntimeException {
     public InvalidHeartTransactionTypeException(String message) {
